structured neatly
