+++ src/main/java/com/titans/travelbooking/controller/StartupPage/IndexController.java
@@ -1,8 +1,7 @@
-package com.titans.travelbooking.controller;
+package com.titans.travelbooking.controller.StartupPage;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
