--- src/main/java/com/titans/travelbooking/dto/AdminRequest.java
@@ -1,23 +1,0 @@
-package com.titans.travelbooking.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Pattern;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class AdminRequest {
-    @NotNull(message = "fullname should not be empty")
-    private String fullName;
-    @Email(message = "enter valid email ")
-    private String username;
-    @Size(min = 8, max = 20, message = "Password must be between 8 and 20 characters")
-    @Pattern(regexp = "^(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=]).*$",
-            message = "Password must contain at least one uppercase letter, one number, and one special")
-    private String password;
-}
