+++ src/main/java/com/titans/travelbooking/service/userservice/IUserService.java
@@ -1,6 +1,6 @@
 package com.titans.travelbooking.service.userservice;
 
-import com.titans.travelbooking.dto.UserRequest;
+import com.titans.travelbooking.validation.UserRequest;
 import com.titans.travelbooking.entity.Users;
 
 public interface IUserService {
