+++ src/main/java/com/titans/travelbooking/service/userservice/UserService.java
@@ -1,6 +1,6 @@
 package com.titans.travelbooking.service.userservice;
 
-import com.titans.travelbooking.dto.UserRequest;
+import com.titans.travelbooking.validation.UserRequest;
 import com.titans.travelbooking.entity.Users;
 import com.titans.travelbooking.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
