+++ src/test/java/com/titans/travelbooking/TravelBookingApplicationTests.java
@@ -6,6 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class TravelBookingApplicationTests {
 
+
+
 	@Test
 	void contextLoads() {
 	}
